Split to separate folder and added Unit Tests
