+++ tools/ni_source_log_processing/.gitignore
@@ -1,0 +1,1 @@
+!test_inputs/*
