+++ tools/ni_source_log_processing/test_inputs/expected-ni_dat_log.csv
@@ -1,0 +1,2 @@
+time,Power in,+12V,+5V,Ox Tank IFM,CCPT 1 IFM,OX Fill TDM,Ox Injector TDM,Fuel Tank TDM,Fuel Injector TDM,Nitrogen Fill TDM,FLintec 100kg Tank Mass,Thrust - CAS BSA-5KLB
+1697637301.0285823,20.127653800287792,24.01044703557456,5.071857711937636,-362.1699972987499,-362.55140486238633,-750.0816949159062,-750.303634661983,-750.0570349441135,-750.0816949159363,-749.8597551698533,-883.661304604026,-13377.245277197455
